requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
beautifulsoup4==4.12.3
pypdf==4.3.1
